refactor: renamed formatDate and update import paths
