+++ src/lib/components/cube/submissionCubeCard.svelte
@@ -1,6 +1,6 @@
 ﻿<script lang="ts">
   import CubeCardSkeleton from "$lib/components/cube/cubeCardSkeleton.svelte";
-  import { formatDate } from "../helper_functions/formatDate.svelte";
+  import { formatDate } from "../helper_functions/formatDate";
   import type { Tables } from "$lib/types/database.types";
 
   interface Props {
