+++ src/routes/(private)/dashboard/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { formatDate } from "$lib/components/helper_functions/formatDate.svelte.js";
+  import { formatDate } from "$lib/components/helper_functions/formatDate.js";
   import Avatar from "$lib/components/user/avatar.svelte";
 
   const { data } = $props();
