+++ src/routes/(public)/explore/cubes/[slug]/reviews/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import StarRating from "$lib/components/rating/starRating.svelte";
   import Avatar from "$lib/components/user/avatar.svelte";
-  import { formatDate } from "$lib/components/helper_functions/formatDate.svelte";
+  import { formatDate } from "$lib/components/helper_functions/formatDate.js";
 
   let { data } = $props();
   let { cube, reviews } = $derived(data);
