+++ src/routes/(public)/user/[username]/achievements/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import type { Enums } from "$lib/types/database.types.js";
-  import { formatDate } from "$lib/components/helper_functions/formatDate.svelte.js";
+  import { formatDate } from "$lib/components/helper_functions/formatDate.js";
   import { getAchievementRarityStyle } from "$lib/components/helper_functions/getAchievementRarityStyle";
   import SearchBar from "$lib/components/misc/searchBar.svelte";
   import Pagination from "$lib/components/misc/pagination.svelte";
